refactor(storage): extract SettingsKeys to dedicated file
Move SettingGroup, SettingContext, SettingKey<T>, SettingsKeys, and
_SettingBoxKey from storage.dart (1067→382 lines) into a new
settings_keys.dart (687 lines). No functional changes.

storage.dart re-exports settings_keys.dart so all existing callers
continue to work without import changes.
